backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
psycopg2-binary
pydantic[email]
python-jose
bcrypt==3.2.2
passlib==1.7.4
aio-pika
python-dotenv
python-multipart
pandas
